edited addresses

+++ LPS35_I2C/LPS35_I2C.h
@@ -10,16 +10,16 @@
 #include "mbed.h"
 
 // SA0 status configuration values.
-#define LPS35_I2C_SA0_HIGH         true
-#define LPS35_I2C_SA0_LOW          false
+#define LPS35_I2C_SA0_HIGH	true
+#define LPS35_I2C_SA0_LOW	false
 
-// Data Rate                                   Pressure / Temperature 
-#define LPS35_I2C_DATARATE_ONESHOT	0x00    // OneShot		OneShot
-#define LPS35_I2C_DATARATE_1HZ		0x01    // 1Hz			1Hz
-#define LPS35_I2C_DATARATE_10HZ		0x02    // 10Hz			10Hz
-#define LPS35_I2C_DATARATE_25HZ		0x03    // 25Hz			25Hz
-#define LPS35_I2C_DATARATE_50HZ		0x04    // 50Hz			50Hz
-#define LPS35_I2C_DATARATE_75HZ		0x05    // 75Hz			75Hz
+// Data Rate   								Pressure / Temperature 
+#define LPS35_I2C_DATARATE_ONESHOT	0x00	// OneShot	OneShot
+#define LPS35_I2C_DATARATE_1HZ		0x01	// 1Hz		1Hz
+#define LPS35_I2C_DATARATE_10HZ		0x02	// 10Hz		10Hz
+#define LPS35_I2C_DATARATE_25HZ		0x03	// 25Hz		25Hz
+#define LPS35_I2C_DATARATE_50HZ		0x04	// 50Hz		50Hz
+#define LPS35_I2C_DATARATE_75HZ		0x05	// 75Hz		75Hz
 
 // I2C Address.
 #define LPS35_I2C_ADDRESS_SA0_HIGH 0xba
@@ -28,26 +28,26 @@
 class LPS35_I2C
 {
 public:
-    LPS35_I2C(PinName sda, PinName scl, bool sa0);
-    ~LPS35_I2C();
+	LPS35_I2C(PinName sda, PinName scl, bool sa0);
+	~LPS35_I2C();
 
-    char whoami();
-    bool isAvailable();
-    
-    void setActive(bool is_active);
-    void setDataRate(char datarate);
+	char whoami();
+	bool isAvailable();
+	
+	void setActive(bool is_active);
+	void setDataRate(char datarate);
 
-    float getPressure();
-    float getTemperature();
+	double getPressure();
+	double getTemperature();
 
-    void _write(char subaddress, char data);
-    char _read(char subaddress);
-    void _read_multibyte(char startsubaddress, char* data, char count);
-        
+	void _write(char subaddress, char data);
+	char _read(char subaddress);
+	void _read_multibyte(char startsubaddress, char* data, char count);
+
 private:
-    I2C _i2c;
-    char _address;
-    char _ctrlreg1;
+	I2C _i2c;
+	char _address;
+	char _ctrlreg1;
 };
 
 #endif /* LPS35_I2C_H */
